+++ src/main/java/com/proyecto_integrador/casa_textil/repositories/UserRepository.java
@@ -5,6 +5,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.proyecto_integrador.casa_textil.entities.Usuario;
  
-public interface UserRepository extends JpaRepository<Usuario, Long> {
+public interface UserRepository extends JpaRepository<Usuario, Integer> {
 	Optional<Usuario> getByEmail(String email);
 }
